Change Detail.js to Detail.tsx
