+++ README.md
@@ -61,3 +61,11 @@ Healmeal 프로젝트
 - 객체 타입 - ITotalMealInfo 처럼 만들기
 - useSelector 오류 : Default -> any로 치환
 - chilren이 JSX 배열일 때 : JSX.Element[];
+
+- hisotry, location 웬만하면 훅 사용할 것
+- 함수의 타입. 웬만하면 useState 함수 넘기지 말것.
+  ```javascript
+  // 인자의 타입 명시
+  setIsMeal: (arg: boolean) => boolean;
+  onClickMeal: () => void;
+  ```
