+++ src/components/Detail/DetailBackground.tsx
